+++ src/Ext/Rules/Body.h
@@ -3,13 +3,13 @@
 #include <CCINIClass.h>
 #include <RulesClass.h>
 #include <GameStrings.h>
-#include <Utilities/Anchor.h>
+
+#include <New/Type/Affiliated/HugeBar.h>
+
 #include <Utilities/Container.h>
 #include <Utilities/Constructs.h>
-#include <Utilities/Template.h>
-#include <Utilities/Enum.h>
-#include <Utilities/TemplateDef.h>
 #include <Utilities/Debug.h>
+#include <Utilities/Enum.h>
 
 class AnimTypeClass;
 class MouseCursor;
@@ -102,59 +102,7 @@ public:
 		Valueable<bool> ShowDesignatorRange;
 		Valueable<bool> IsVoiceCreatedGlobal;
 
-		class HugeBarData
-		{
-		public:
-
-			Valueable<double> HugeBar_RectWidthPercentage;
-			Valueable<Point2D> HugeBar_RectWH;
-			Damageable<ColorStruct> HugeBar_Pips_Color1;
-			Damageable<ColorStruct> HugeBar_Pips_Color2;
-
-			Valueable<SHPStruct*> HugeBar_Shape;
-			Valueable<SHPStruct*> HugeBar_Pips_Shape;
-			CustomPalette HugeBar_Palette;
-			CustomPalette HugeBar_Pips_Palette;
-			Damageable<int> HugeBar_Frame;
-			Damageable<int> HugeBar_Pips_Frame;
-			Valueable<int> HugeBar_Pips_Spacing;
-
-			Valueable<Point2D> HugeBar_Offset;
-			Nullable<Point2D> HugeBar_Pips_Offset;
-			Valueable<int> HugeBar_Pips_Num;
-
-			Damageable<ColorStruct> Value_Text_Color;
-
-			Valueable<SHPStruct*> Value_Shape;
-			CustomPalette Value_Palette;
-			Valueable<int> Value_Num_BaseFrame;
-			Valueable<int> Value_Sign_BaseFrame;
-			Valueable<int> Value_Shape_Spacing;
-
-			Valueable<bool> DisplayValue;
-			Valueable<bool> Value_Percentage;
-			Valueable<Point2D> Value_Offset;
-			Anchor Anchor;
-			DisplayInfoType InfoType;
-
-			Valueable<bool> VisibleToHouses_Observer;
-			Valueable<AffectedHouse> VisibleToHouses;
-
-			HugeBarData() = default;
-			HugeBarData(DisplayInfoType infoType);
-
-			void LoadFromINI(CCINIClass* pINI);
-
-			bool Load(PhobosStreamReader& stm, bool registerForChange);
-			bool Save(PhobosStreamWriter& stm) const;
-
-		private:
-
-			template <typename T>
-			bool Serialize(T& stm);
-		};
-
-		std::vector<std::unique_ptr<HugeBarData>> HugeBar_Config;
+		std::vector<std::unique_ptr<HugeBar>> HugeBar_Config;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
