+++ src/Ext/Techno/Body.cpp
@@ -8,13 +8,13 @@
 
 #include <Ext/House/Body.h>
 
+#include <New/Type/Affiliated/HugeBar.h>
+
 #include <Utilities/AresFunctions.h>
 #include <Utilities/EnumFunctions.h>
 #include <Utilities/ShapeTextPrinter.h>
 
 TechnoExt::ExtContainer TechnoExt::ExtMap;
-
-std::vector<std::pair<int, TechnoClass*>> TechnoExt::Techno_HugeBar;
 
 TechnoExt::ExtData::~ExtData()
 {
@@ -631,360 +631,6 @@ bool TechnoExt::IsTypeImmune(TechnoClass* pThis, TechnoClass* pSource)
 	return false;
 }
 
-void TechnoExt::InitializeHugeBar(TechnoClass* pThis)
-{
-	auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
-
-	if (pTypeExt->HugeBar)
-	{
-		auto itFinded = Techno_HugeBar.cend();
-
-		for (auto it = Techno_HugeBar.cbegin(); it != Techno_HugeBar.cend(); it++)
-		{
-			if (it->second == pThis)
-			{
-				itFinded = it;
-				break;
-			}
-		}
-
-		if (itFinded != Techno_HugeBar.cend())
-		{
-			Techno_HugeBar.emplace_back(pTypeExt->HugeBar_Priority, pThis);
-		}
-	}
-}
-
-void TechnoExt::ProcessHugeBar()
-{
-	if (Techno_HugeBar.empty())
-		return;
-
-	const auto& configs = RulesExt::Global()->HugeBar_Config;
-
-	for (size_t i = 0; i < configs.size(); i++)
-	{
-		TechnoClass* pTechno = nullptr;
-		int priority = MIN(int);
-
-		for (const auto& item : Techno_HugeBar)
-		{
-			TechnoClass* pTmpTechno = item.second;
-
-			if (HouseClass::IsCurrentPlayerObserver() && !configs[i]->VisibleToHouses_Observer)
-				continue;
-
-			if (!HouseClass::IsCurrentPlayerObserver()
-				&& !EnumFunctions::CanTargetHouse(
-					configs[i]->VisibleToHouses,
-					pTmpTechno->GetOwningHouse(),
-					HouseClass::CurrentPlayer))
-				continue;
-
-			const auto pTmpTechnoTypeExt = TechnoTypeExt::ExtMap.Find(pTmpTechno->GetTechnoType());
-
-			if (pTmpTechnoTypeExt->HugeBar_Priority > priority)
-			{
-				priority = pTmpTechnoTypeExt->HugeBar_Priority;
-				pTechno = pTmpTechno;
-			}
-		}
-
-		if (pTechno == nullptr)
-			continue;
-
-		int iCurrent = -1;
-		int iMax = -1;
-		GetValuesForDisplay(pTechno, configs[i]->InfoType, iCurrent, iMax);
-
-		if (iCurrent != -1)
-			DrawHugeBar(configs[i].get(), iCurrent, iMax);
-	}
-}
-
-void TechnoExt::DrawHugeBar(RulesExt::ExtData::HugeBarData* pConfig, int iCurrent, int iMax)
-{
-	double ratio = static_cast<double>(iCurrent) / iMax;
-	int iPipNumber = std::max(static_cast<int>(ratio * pConfig->HugeBar_Pips_Num), (iCurrent == 0 ? 0 : 1));
-	RectangleStruct rectDraw = DSurface::Composite->GetRect();
-	rectDraw.Height -= 32;
-	Point2D posDraw = pConfig->HugeBar_Offset.Get() + pConfig->Anchor.OffsetPosition(rectDraw);
-	Point2D posDrawValue = posDraw;
-	RectangleStruct rBound = std::move(DSurface::Composite->GetRect());
-
-	if (pConfig->HugeBar_Shape != nullptr
-		&& pConfig->HugeBar_Pips_Shape != nullptr
-		&& pConfig->HugeBar_Frame.Get(ratio) >= 0
-		&& pConfig->HugeBar_Pips_Frame.Get(ratio) >= 0)
-	{
-		SHPStruct* pShp_Bar = pConfig->HugeBar_Shape;
-		ConvertClass* pPal_Bar = pConfig->HugeBar_Palette.GetOrDefaultConvert(FileSystem::PALETTE_PAL);
-		SHPStruct* pShp_Pips = pConfig->HugeBar_Pips_Shape;
-		ConvertClass* pPal_Pips = pConfig->HugeBar_Pips_Palette.GetOrDefaultConvert(FileSystem::PALETTE_PAL);
-		int iPipFrame = pConfig->HugeBar_Pips_Frame.Get(ratio);
-
-		switch (pConfig->Anchor.Horizontal)
-		{
-		case HorizontalPosition::Left:
-			posDrawValue.X += pShp_Bar->Width / 2;
-			break;
-
-		case HorizontalPosition::Center:
-			posDrawValue.X = posDraw.X;
-			posDraw.X -= pShp_Bar->Width / 2;
-			break;
-
-		case HorizontalPosition::Right:
-			posDraw.X -= pShp_Bar->Width;
-			posDrawValue.X -= pShp_Bar->Width / 2;
-			break;
-
-		default:
-			break;
-		}
-
-		switch (pConfig->Anchor.Vertical)
-		{
-		case VerticalPosition::Top:
-			posDrawValue.Y += pShp_Bar->Height;
-			break;
-
-		case VerticalPosition::Center:
-			posDraw.Y -= pShp_Bar->Height / 2;
-			posDrawValue.Y += pShp_Bar->Height;
-			break;
-
-		case VerticalPosition::Bottom:
-			posDraw.Y -= pShp_Bar->Height;
-			posDrawValue.Y -= pShp_Bar->Height;
-			break;
-
-		default:
-			break;
-		}
-
-		DSurface::Composite->DrawSHP
-		(
-			pPal_Bar,
-			pShp_Bar,
-			pConfig->HugeBar_Frame.Get(ratio),
-			&posDraw,
-			&rBound,
-			BlitterFlags::None,
-			0,
-			0,
-			ZGradient::Ground,
-			1000,
-			0,
-			nullptr,
-			0,
-			0,
-			0
-		);
-
-		posDraw += pConfig->HugeBar_Pips_Offset.Get();
-
-		for (int i = 0; i < iPipNumber; i++)
-		{
-			DSurface::Composite->DrawSHP
-			(
-				pPal_Pips,
-				pShp_Pips,
-				iPipFrame,
-				&posDraw,
-				&rBound,
-				BlitterFlags::None,
-				0,
-				0,
-				ZGradient::Ground,
-				1000,
-				0,
-				nullptr,
-				0,
-				0,
-				0
-			);
-
-			posDraw.X += pConfig->HugeBar_Pips_Spacing;
-		}
-	}
-	else
-	{
-		COLORREF color1 = Drawing::RGB_To_Int(pConfig->HugeBar_Pips_Color1.Get(ratio));
-		COLORREF color2 = Drawing::RGB_To_Int(pConfig->HugeBar_Pips_Color2.Get(ratio));
-		Vector2D<int> rectWH = pConfig->HugeBar_RectWH;
-
-		if (rectWH.X < 0)
-		{
-			rectWH.X = static_cast<int>(pConfig->HugeBar_RectWidthPercentage * rBound.Width);
-			// make sure width is even
-			rectWH.X += rectWH.X & 1;
-		}
-
-		switch (pConfig->Anchor.Horizontal)
-		{
-		case HorizontalPosition::Left:
-			posDrawValue.X += rectWH.X / 2;
-			break;
-
-		case HorizontalPosition::Center:
-			posDrawValue.X = posDraw.X;
-			posDraw.X -= rectWH.X / 2;
-			break;
-
-		case HorizontalPosition::Right:
-			posDraw.X -= rectWH.X;
-			posDrawValue.X -= rectWH.X / 2;
-			break;
-
-		default:
-			break;
-		}
-
-		switch (pConfig->Anchor.Vertical)
-		{
-		case VerticalPosition::Top:
-			posDrawValue.Y += rectWH.Y;
-			break;
-
-		case VerticalPosition::Center:
-			posDraw.Y -= rectWH.Y / 2;
-			posDrawValue.Y += rectWH.Y;
-			break;
-
-		case VerticalPosition::Bottom:
-			posDraw.Y -= rectWH.Y;
-			posDrawValue.Y -= rectWH.Y;
-			break;
-
-		default:
-			break;
-		}
-
-		RectangleStruct rect = { posDraw.X, posDraw.Y, rectWH.X, rectWH.Y };
-		DSurface::Composite->DrawRect(&rect, color2);
-		int iPipWidth = 0;
-		int iPipHeight = 0;
-		int iPipTotal = pConfig->HugeBar_Pips_Num;
-
-		if (pConfig->HugeBar_Pips_Offset.isset())
-		{
-			Point2D offset = pConfig->HugeBar_Pips_Offset.Get();
-			posDraw += offset;
-			//center
-			iPipWidth = (rectWH.X - offset.X * 2) / iPipTotal;
-			iPipHeight = rectWH.Y - offset.Y * 2;
-		}
-		else
-		{
-			// default has 5 interval between border and pips at least
-			const int iInterval = 5;
-			iPipWidth = (rectWH.X - iInterval * 2) / iPipTotal;
-			iPipHeight = rectWH.Y - iInterval * 2;
-			posDraw.X += (rectWH.X - iPipTotal * iPipWidth) / 2;
-			posDraw.Y += (rectWH.Y - iPipHeight) / 2;
-		}
-
-		if (iPipWidth <= 0 || iPipHeight <= 0)
-			return;
-
-		// Color1 75% Color2 25%, simulated healthbar
-		int iPipColor1Width = std::max(static_cast<int>(iPipWidth * 0.75), std::min(3, iPipWidth));
-		int iPipColor2Width = iPipWidth - iPipColor1Width;
-		rect = { posDraw.X, posDraw.Y, iPipColor1Width , iPipHeight };
-
-		for (int i = 0; i < iPipNumber; i++)
-		{
-			DSurface::Composite->FillRect(&rect, color1);
-			rect.X += iPipColor1Width;
-			rect.Width = iPipColor2Width;
-			DSurface::Composite->FillRect(&rect, color2);
-			rect.X += iPipColor2Width;
-			rect.Width = iPipColor1Width;
-		}
-	}
-
-	HugeBar_DrawValue(pConfig, posDrawValue, iCurrent, iMax);
-}
-
-void TechnoExt::HugeBar_DrawValue(RulesExt::ExtData::HugeBarData* pConfig, Point2D& posDraw, int iCurrent, int iMax)
-{
-	RectangleStruct rBound = std::move(DSurface::Composite->GetRect());
-	double ratio = static_cast<double>(iCurrent) / iMax;
-	posDraw += pConfig->Value_Offset;
-
-	if (pConfig->Value_Shape != nullptr)
-	{
-		SHPStruct* pShp = pConfig->Value_Shape;
-		ConvertClass* pPal = pConfig->Value_Palette.GetOrDefaultConvert(FileSystem::PALETTE_PAL);
-
-		if (pConfig->Anchor.Vertical == VerticalPosition::Bottom)
-			posDraw.Y -= pShp->Height * (static_cast<int>(pConfig->InfoType) + 1);
-		else
-			posDraw.Y += pShp->Height * static_cast<int>(pConfig->InfoType);
-
-		std::string text;
-
-		if (pConfig->Value_Percentage)
-			text = GeneralUtils::IntToDigits(static_cast<int>(ratio * 100));
-		else
-			text = GeneralUtils::IntToDigits(iCurrent) + "/" + GeneralUtils::IntToDigits(iMax);
-
-		int iNumBaseFrame = pConfig->Value_Num_BaseFrame;
-		int iSignBaseFrame = pConfig->Value_Sign_BaseFrame;
-
-		if (ratio <= RulesClass::Instance->ConditionYellow)
-		{
-			// number 0-9
-			iNumBaseFrame += 10;
-			// sign /%
-			iSignBaseFrame += 2;
-		}
-
-		if (ratio <= RulesClass::Instance->ConditionRed)
-		{
-			iNumBaseFrame += 10;
-			iSignBaseFrame += 2;
-		}
-
-		posDraw.X -= text.length() * pConfig->Value_Shape_Spacing / 2;
-
-		ShapeTextPrintData printData
-		(
-			pShp,
-			pPal,
-			iNumBaseFrame,
-			iSignBaseFrame,
-			Point2D({ pConfig->Value_Shape_Spacing, 0 })
-		);
-		ShapeTextPrinter::PrintShape(text.c_str(), printData, posDraw, rBound, DSurface::Composite);
-	}
-	else
-	{
-		const int iTextHeight = 15;
-
-		if (pConfig->Anchor.Vertical == VerticalPosition::Bottom)
-			posDraw.Y -= iTextHeight * (static_cast<int>(pConfig->InfoType) + 1);
-		else
-			posDraw.Y += iTextHeight * static_cast<int>(pConfig->InfoType);
-
-		wchar_t text[0x16] = L"";
-
-		if (pConfig->Value_Percentage)
-		{
-			swprintf_s(text, L"%d", static_cast<int>(ratio * 100));
-			wcscat_s(text, L"%%");
-		}
-		else
-		{
-			swprintf_s(text, L"%d/%d", iCurrent, iMax);
-		}
-
-		COLORREF color = Drawing::RGB_To_Int(pConfig->Value_Text_Color.Get(ratio));
-		DSurface::Composite->DrawTextA(text, &rBound, &posDraw, color, COLOR_BLACK, TextPrintType::Center);
-	}
-}
-
 // =============================
 // load / save
 
@@ -1024,14 +670,12 @@ void TechnoExt::ExtData::SaveToStream(PhobosStreamWriter& Stm)
 bool TechnoExt::LoadGlobals(PhobosStreamReader& Stm)
 {
 	return Stm
-		.Process(Techno_HugeBar)
 		.Success();
 }
 
 bool TechnoExt::SaveGlobals(PhobosStreamWriter& Stm)
 {
 	return Stm
-		.Process(Techno_HugeBar)
 		.Success();
 }
