+++ src/Utilities/Anchor.h
@@ -1,8 +1,7 @@
 #pragma once
-#include <Utilities/Enum.h>
-#include <Utilities/TemplateDef.h>
 
-#include <YRPPCore.h>
+#include <Utilities/Enum.h>
+#include <Utilities/Template.h>
 
 // Helper class to get anchor points on an arbitrary rectangle or a parallelogram
 class Anchor
